Grading
Good work,

You really got me with that heatmap question in class Thursday. Really these are essentially two of the same layers. However because of that I'm going to dock you on UI since you used the checkbox and not the radio button (groups instead of base). Ideally you'd only want folks to see either the points or the heatmap, not both at once. Also, that small issue with the datatable I mentioned in the comments.